// fe-demo/src/interfaces/Command.ts
export interface CommandFormState {
  so_van_ban: string;
  ngay_ban_hanh: string;
  ngay_nhan: string;
  don_vi_gui: string;
  loai_van_ban: string;
  noi_dung: string;
  lanh_dao_id: string;
}

export interface CommandFilterValues {
  loai_van_ban: string;
  don_vi_gui: string;
  ngay_nhan: string;
}